+++ src/main/java/nus/iss/se/magicbag/interfacemethods/CartInterface.java
@@ -7,13 +7,14 @@ import nus.iss.se.magicbag.entity.CartItem;
 
 
 public interface CartInterface {
-    Cart createCart(long userId);
-    Cart getActiveCart(long userId);
-    Cart addItemToCart(long userId, long magicbagId, int quantity);
-    Cart updateItemQuantityInCart(long userId, long magicbagId, int newQuantity);
-    Cart removeItemFromCart(long userId, long magicbagId);
-    List<CartItem> getCartItems(long userId);
-    Cart clearCart(long userId);
-    double getTotal(long userId);
-    List<CartItem> getCartItemsByMagicBagId(long id);
+    Cart createCart(Integer userId);
+    Cart getActiveCart(Integer userId);
+    Cart addItemToCart(Integer userId, Integer magicbagId, int quantity);
+    Cart updateItemQuantityInCart(Integer userId, Integer magicbagId, int newQuantity);
+    Cart removeItemFromCart(Integer userId, Integer magicbagId);
+    List<CartItem> getCartItems(Integer userId);
+    Cart clearCart(Integer userId);
+    double getTotal(Integer userId);
+    List<CartItem> getCartItemsByMagicBagId(Integer id);
 }
+
